backend/requirements.txt
asgiref==3.9.1
asttokens==3.0.0
attrs==20.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
comm==0.2.3
cryptography==45.0.6
debugpy==1.8.15
decorator==5.2.1
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
executing==2.2.0
idna==3.10
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jellyfish==1.2.0
jupyter_client==8.6.3
jupyter_core==5.8.1
lib50==3.1.4
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311; sys_platform == "win32"
PyYAML==6.0.2
pyzmq==27.0.1
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
submit50==3.2.1
termcolor==1.1.0
tornado==6.5.1
traitlets==5.14.3
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
whitenoise==6.10.0
